Refactor out functions into their own file as they are large
